+++ controllers/medication_statement.js
@@ -1,8 +1,9 @@
 const medState = require("../models/MedicationStatement");
 const Patients = require("../models/patientResource");
 const Mongoose = require("mongoose");
-
 const { result_controller } = require("../middleware");
+
+const { getPatientById } = require("./patients");
 
 const getAllMedState = async () => {
   try {
@@ -28,6 +29,50 @@ const getMedStateById = async (id) => {
   }
 };
 
+const getMedStateByPatientId = async (patientId) => {
+  try {
+    let checkPatient = await getPatientById(patientId);
+    if (checkPatient.data != null) {
+      let medStateData = await Patients.findById(patientId).populate({
+        path: "extension.medicationStatment",
+        model: "MedicationStatement",
+      });
+      return result_controller("OK", medStateData.extension.medicationStatment);
+    } else {
+      return result_controller("ERROR patient not found", null);
+    }
+  } catch (error) {
+    console.error(error.message);
+    return result_controller("ERROR", null);
+  }
+};
+
+const getSingleMedStateByPatientId = async (patientId, medStateId) => {
+  try {
+    let checkPatient = await getPatientById(patientId);
+    if (checkPatient.data != null) {
+      let medStateData = await Patients.findById(patientId).populate({
+        path: "extension.medicationStatment",
+        model: "MedicationStatement",
+        match: { _id: medStateId },
+      });
+      if (medStateData.extension.medicationStatment.length != 0) {
+        return result_controller(
+          "OK",
+          medStateData.extension.medicationStatment
+        );
+      } else {
+        return result_controller("ERROR medical statement not found", null);
+      }
+    } else {
+      return result_controller("ERROR patient not found", null);
+    }
+  } catch (error) {
+    console.error(error.message);
+    return result_controller("ERROR", null);
+  }
+};
+
 const updatePatientMedState = async (patientId, medStateId) => {
   try {
     const result = await Patients.findByIdAndUpdate(
@@ -46,12 +91,20 @@ const updatePatientMedState = async (patientId, medStateId) => {
 
 const createMedState = async (newData) => {
   try {
-    const result = await medState.create(newData);
-    //update patient resource
     let patientId = newData.subject;
-    let medStateId = result.id;
-    const updatePatient = await updatePatientMedState(patientId, medStateId);
-    return result_controller("OK", result);
+    let patientObjectId = Mongoose.Types.ObjectId(patientId);
+    newData.subject = patientObjectId;
+    let checkPatient = await getPatientById(patientId);
+
+    if (checkPatient.data != null) {
+      const result = await medState.create(newData);
+      //update patient resource
+      let medStateId = result.id;
+      const updatePatient = await updatePatientMedState(patientId, medStateId);
+      return result_controller("OK", result);
+    } else {
+      return result_controller("ERROR Patient Id not found", null);
+    }
   } catch (error) {
     console.error(error);
     return result_controller("ERROR", null);
@@ -92,15 +145,20 @@ const removePatientMedState = async (patientId, medStateId) => {
 
 const deleteMedStateById = async (patientId, medStateId) => {
   try {
-    const result = await removePatientMedState(patientId, medStateId);
-    const deletedData = await medState.findByIdAndRemove(medStateId);
-    if (deletedData && result) {
-      return result_controller("OK", result);
+    let checkPatient = await getPatientById(patientId);
+    if (checkPatient.data != null) {
+      const deletedData = await medState.findByIdAndRemove(medStateId);
+      if (deletedData != null) {
+        const result = await removePatientMedState(patientId, medStateId);
+        return result_controller("OK", deletedData);
+      } else {
+        return result_controller(
+          "ERROR, medication state data not found",
+          null
+        );
+      }
     } else {
-      return result_controller(
-        "ERROR, medical statement or patient data not found",
-        null
-      );
+      return result_controller("ERROR, patient data not found", null);
     }
   } catch (error) {
     console.error(error);
@@ -111,6 +169,8 @@ const deleteMedStateById = async (patientId, medStateId) => {
 module.exports = {
   getAllMedState,
   getMedStateById,
+  getMedStateByPatientId,
+  getSingleMedStateByPatientId,
   createMedState,
   updateMedStateById,
   deleteMedStateById,
